# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ImmoPro API",
        "description": "Documentation interactive de l'API ImmoPro — plateforme immobilière.\n\n## Authentification\nLa majorité des endpoints nécessitent un token Bearer (Laravel Sanctum).\nConnectez-vous via **POST /api/login** (Admin/Agent) ou **POST /api/client/login** (Client),\npuis cliquez sur le bouton **Authorize** et collez votre token.\n\n## Rôles\n- **admin** : accès complet à l'administration\n- **agent** : gestion des biens assignés et des visites\n- **client** : propriétaires / acheteurs / locataires",
        "contact": {
            "name": "Support ImmoPro",
            "email": "support@immopro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal de l'API ImmoPro"
        }
    ],
    "paths": {
...
        "/biens": {
            "get": {
                "tags": [
                    "Biens Immobiliers (Public)"
                ],
                "summary": "Lister les biens publiés",
                "description": "Retourne la liste paginée des biens immobiliers publiés. Accessible sans authentification.",
                "operationId": "listBiensPublics",
                "parameters": [
                    {
                        "name": "type_bien",
                        "in": "query",
                        "description": "Type de bien",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_transaction",
                        "in": "query",
                        "description": "Type de transaction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "vente",
                                "location",
                                "colocation"
                            ]
                        }
                    },
                    {
                        "name": "prix_min",
                        "in": "query",
                        "description": "Prix minimum",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "prix_max",
                        "in": "query",
                        "description": "Prix maximum",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "surface_min",
                        "in": "query",
                        "description": "Surface min (m²)",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "ville",
                        "in": "query",
                        "description": "Ville / Adresse",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche textuelle",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "prix_asc",
                                "prix_desc",
                                "date_desc",
                                "surface_desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Filtres invalides"
                    }
                }
            }
        },
        "/biens/{id}": {
            "get": {
                "tags": [
                    "Biens Immobiliers (Public)"
                ],
                "summary": "Détail d'un bien publié",
                "description": "Retourne le détail complet d'un bien immobilier publié.",
                "operationId": "showBienPublic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du bien récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bien introuvable"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "title": "Métadonnées de pagination",
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 120
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 8
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu après connexion. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}